1.Merge DDR3/DDR4/DDR_FPGA init file to firmware and clear up the code. 2.Complete the acu_ddr test cases. 3.Complete PLL modules function. 4.Complete PVT init clock warning issue. 5.Complete Efuse module function. 6.Modify test case printf info SUCCESS to PASSED. 7.Modify KEIL build defines in readme.txt. 8.Move peie ELBI functions to acu_pcie.c. 9.Modify simulation printf register to 0x2001FFFC.

+++ src/driver/src/acu_ddr.c
@@ -218,9 +218,9 @@ void DDR_ReadMem(uint8_t *buf, uint32_t addr, uint32_t size, uint8_t channel)
     }
 }
 
-void DDR_InterleaveWriteMem(uint8_t *buf, uint64_t addr, uint32_t size)
+void DDR_InterleaveWriteMem(uint8_t *buf, uint32_t addr, uint32_t size)
 {
-    uint64_t PageHeadAddr = 0;
+    uint32_t PageHeadAddr = 0;
     uint8_t PageProgAddr = 0;
     uint32_t RowNum = 0, LineNum = 0;
     uint32_t PageNum = 0, PageMod = 0;
@@ -283,9 +283,9 @@ void DDR_InterleaveWriteMem(uint8_t *buf, uint64_t addr, uint32_t size)
     }
 }
 
-void DDR_InterleaveReadMem(uint8_t *buf, uint64_t addr, uint32_t size)
+void DDR_InterleaveReadMem(uint8_t *buf, uint32_t addr, uint32_t size)
 {
-    uint64_t PageHeadAddr = 0;
+    uint32_t PageHeadAddr = 0;
     uint8_t PageProgAddr = 0;
     uint32_t RowNum = 0, LineNum = 0;
     uint32_t PageNum = 0, PageMod = 0;
